Add write_text check:
Add check for `Path(x).write_text()`. In implementing this check, I realized
that the `read_text` check doesnt actually check that the resource created
in the `with` is the same one being written to, so I fixed that. I also
renamed the `read_x` check to `read_text`. Similar to the `startswith` check,
I am just going to chose a single name for the check, which might encompass
multiple variations.