Add check for replacing `x == y or x == z` with `x in (y, z)`:
This check feels a bit finicky, since any node which has a similar string
representation as `x` is allowed. This might result in some false negatives,
since we are relying on the `__str__()` method of the particular node.

Also, this check can't handle multiple `or` exprs chained together. This might
be hard to fix, since the nodes are parsed in a recursive fasion, potentially
resulting in multiple errors for the same expr.

Also, this check fails to notice if calling `x` has side-effects, or if `y`
or `z` depend on being short-circuted. In either case, this check should be
ignored.