Add list `.extend()` check:
Small bug/annoyance with this implementation is that the errors are not sorted
by source line/column. I will make another commit for this later, since it
has always existed, but is just now being noticed.

Note that this check will not be able to check 2 append statements which have a
non-dependant statement (ie, a print call) inbetween. That kind of a check
would be hard to do, and very error prone. This should suffice for most cases.