Add check for lambdas that can use the operator module:
This basically just finds lambdas like `lambda x, y: x + y`, and suggests it
be replaced with its corresponding `operator` function (in this case, `add`).
This check applies to normal functions as well.

There are a lot of other functions in the `operator` module which I didn't
care to implement, specifically the in place operators (`+=`), and slice
operators, and so on.