Add checks for explicitly using default value of a function:
This ended up being a whole lot more complicated then I thought it was going to
be. For starters, there are multiple ways that you can call a function, and many
ways that Mypy represents a callable in it's AST tree: Some are functions, some
are overloaded functions, some are class objects (such as `int`), and some I am
probably missing.

A hack that I had to do is "inject" the defaults for certain functions in the
stdlib, such as `input` and `round`. These functions have well documented
defaults, but the actual code does not specify these defaults, and instead
relies on a lot of generic magic. I only added a few just to get things rolling,
and more can be added as needed.

There shouldn't be any outstanding issues with this, though I have yet to test
Refurb on a sizable codebase. With all the checks I have been adding,
performance testing might be something to look into, since I have no idea how
fast/slow these checks are.

This check has already gotten big enough, and there are still things I needed to
do, such as checking the constructors (ie, `int`).