Improve pattern matching logic for `or` expression checks:
Certain `or` expressions require more checks to catch all scenarios, such as
nested `or` expressions. Since this extraction technique will be useful for all
checks which are pattern matching binary expressions, I made it into a
funciton. The only downside with my approach is that there is an extra `if`
check that needs to be done before you can pattern match, since there is a bug
in Mypy which causes a crash when linting (though it runs just fine). It is in
master, and will be pushed to pypi soon.