Fix some bugs in the fstring format check:
Using `int(x)` inside of an f-string can be valid in some cases. For example,
when `x` is " 123 ", the spaces are stripped, and you are left with `123`. But,
if you remove the surrounding `int()` call, the spaces remain.

Also, only single parameter functions are allowed now. Before, things like
`str(b"abc", encoding="utf8")` where allowed, which is not (always) going to
be the same.