Add "no unnecessary cast" check:
This check creates some extra noise in some of the other test cases, and so
it would be nice to use an inline comment to disable that. I can't seem to
find an easy way to do it (from within mypy), so I might need to open and
read the files individually, which would not be ideal.