Fix bug with simplify return check:
One of the requirements for an `else` or `case _` block being able to be
simplified is that all the above blocks must end with a `return`, otherwise
simplifying the return would change the behaviour. This change might be too
strict though, since it does not allow recursively check that all paths
contain a return, so there might be some false negatives, which is fine for
now.