Create "gen" subcommand for generating new checks:
This adds a new subcommand for quickly creating new checks. First you select
the type of AST node you want to visit, and then you type the file path of
where you want the new check to go. This will create new folders if needed,
but will not add any `__init__.py` files (this will be fixed later). Also,
the code will not be auto-incremented, since there is no good way of finding
the next number aside from loading all the checks (which might need to be
done anyways).