Fix `cast()` expressions causing 2 errors to be emitted:
The comment in the source code explains this, but basically there is an
`analyzed` field in CallExpr's, and traversing into them causes the same
CallExpr node to be hit again. This is not what I want, so I changed that
particular visitor to not visit the CastExpr. This might cause issues if
you explicitly want to visit the CastExpr, but I will fix that later if
it comes up.