Add ability to parse config settings from `pyproject.toml` file:
TODO: Document these changes. This basically allows you to store the `load` and
`ignore` values in a config file. The rest of the CLI args do not make sense
to be a config file, which is more suited for long-term, consistent settings.
For example, `--debug` dont make sense to be in the config.

I chose to pull from the `pyproject.toml` file because it is the newer, more
modern standard for python projects. Also, existing linters such as mypy and
black already support the pyproject convention, so having all your tool configs
in one place will be very convinent.