Make error code system more robust:
Previously, error codes where stored as ints, which meant that custom error
codes would conflict with the built-in ones. Now there is a dedicated ErrorCode
class for storing the prefix and id of the error.

This commit also does a lot of cleanup relating to the module loading system.
Since I noticed I had to update both the explain module loader and the check
module loader, I decided to just isolate the code into one place.