Add ability to ignore errors with inline comments:
This is a very "dumb" implementation, meaning that it doesnt parse the code to
look for comments, it merely checks if the line in question ends in a comment.
This will be good for most use cases, but one instance where it will break down
is in the following (contrived) example:

```python
print(f"""{'some_example_code'} # noqa
""")
```

The `noqa` comment is not a comment, but part of a multi-line f-string. This
sort of scenario should be rare, but something to keep in mind. Perhaps there
is a more common scenario that I am missing, but regardless, what is currently
in place should be fine.