Bump mypy:
Sadly, https://github.com/python/mypy/pull/13728 was not merged into the 0.981
release, meaning we still have the weird None check when using the `extract_binary_oper`
function. I could just use the master branch instead of the pypi release of
mypy, but those don't have wheels built for it, meaning it would (probably) be
a lot harder to install for the end user.