Fix `Path.read_text()` false positive:
`Path.read_text()` will read the entire contents of a file into a variable,
whereas `f.read(x)` will only read `x` characters/bytes. This check did not
check for that use case, hence the false positive.

In addition to that argument, Refurb also did not check the other args to `open()`,
some of which are not supported in `read_text()`. Only the "encoding"/"errors"
options are supported. Now if an unsupported option is passed, say "buffering",
no error will be emitted.

I also made the display a bit nicer, indicating when there are/are not
arguments being passed to open/read_text.