Require len() comparison check to only operate within `if` statements:
Previously, this check picked up on cases which might cause a different value
or type to be returned if the suggestion was blindly applied. The assumption
was that you would be doing this in a boolean context, such as an `if` block,
which would handle the truthy value correctly. Now this check only checks code
inside of an `if` statement (including match guards, "if" sections of list
generators, and so on).