Simplify error code naming system:
Currently, the error code class for a given check has to have an arbitrary
name which must start with "Error", but cannot be "ErrorCode" or "Error". Given
that the name of the error class is not used/displayed anywhere, and doesn't
provide any extra context, having to give them a name (similar to the filename)
doesn't make a lot of sense. I've updated all checks to be called "ErrorInfo",
making all the checks more consistent (similar to the name of the "check"
functions). At some point I might make it required for these error info classes
to be named "ErrorInfo", but that would be a breaking change.

If we do want to "name" the checks, an explicit "name" field in the class would
be a better option, and better then relying on the classname.

As an aside, there should probably be warnings that go off if you have a check
function or error class which is not the right type, or for whatever reason, your
check cannot be loaded.