Fix incorrect error message for `dict`/`list` for FURB123: (#36)
`dict` and `list` create shadow copies, which might be intentional, and
following the suggestion could result in a bug. This fixes it, and instead
suggests using `.copy()` instead. There are plenty of ways to make shallow
copies of dicts and lists, but this is built into the dict/list objects,
and doesn't rely on newer language features.

Some people will still like to write `dict(x)`, but still might want to
enable this check. At some point, I would like to add the ability to disable
errors with more granular control (if the check allows it). For example,
you could globally disable `FURB123:dict` to opt-out of error checking FURB123
on dict types.