Fix operator precedence not being upheld:
As it turns out I was checking that the operator was the same as the requested
operator, but not fully. Basically, I needed to bail out if the OpExpr is not
the same type, otherwise it would hit the next line, which is an Expression,
which OpExpr derives from, and incorrectly returning it.