Partially fix variable type issues:
Apparently Mypy does not update the type of a variable when it is assigned to
a new type, it will just keep the old type. There is a flag that allows for
redefinitions of variables (as a different type), but it will only allow the
type to be changed if the prior variable was read from.

This should fix common use cases, but not all of them. If I cannot seem to get
Mypy working the way I need it to, I might have to fork it. This might be
preferable anyways, since there are some design changes which I would like to
make which are hard to do with the official Mypy repo.