Add --config-file flag:
This allows for changing the default config file that Refurb loads.

In the future, Refurb will try to pull the `pyproject.toml` file from the root
of the project structure, and fall back to the current directory if none is
found. For now, this will work in case the config file is stored in a
completely different location.

Note that even if the config file is not named `pyproject.toml`, it will
still have to follow the same structure, including the `[tool.refurb]` section.