Add ability to disable checks (#73):
The ability to enable checks was added in 1ede2fe0. This is a continuation of
that, allowing you to disable checks which are enabled by default.

In the future, the enable/disable functionality will need to be improved upon,
since currently you can only enable/disable error codes, but in the future I
plan on being able to the same for categories.