Add "set discard" check:
When removing keys from a `set`, you can use `.remove()`, which assumes the
key already exists. If you just want to remove the key regardless of
whether it is already in the set or not, use `.discard()` instead.

One potential case that this check does not pick up on is the following:

```python
try:
    nums.remove(123)
except KeyError:
    pass
```

Since `.remove()` will throw a `KeyError` if the key doesn't exist, a valid
way of silencing this would be the above code. This is identical to the "with
suppress" check, and I don't like the interaction that all of these checks
are having with it. This also means that the following code should be checked:

```python
with suppress(KeyError):
    nums.remove(123)
```

Though adding 3 different styles for the same check is a lot of work, and
requires additional maintainence.

Also, this check (and many others) do not pick up on member expressions, such
as `x.y.z.remove(123)`. This will be addressed soon.