Bump packages:
I removed the poetry lock file and the requirements.txt file, since this is
a library, and the exact versions of the installed packages do not matter for
the end user, just that the correct range is in place.

In addition, proper version ranges have been added for `mypy` and `type_extentions`.

Closes #40.