Fix false positive in FURB125:
It is common to use `return` or `pass` as an early exit in a match statement,
and Refurb was treating these as errors. Although Refurb is correct in that the
return is not needed (`pass` would work as well), removing the case entirely
would cause an error.

The updated logic is as follows:

* Wildcards will error if there is a trailing return in the body.

* Non-wildcard patterns will error if the length of the body is more then one.

* Non-wildcard patterns with just a `return` statement will not be an error.