Add `--python-version` flag:
This allows for specifying which version of Python you are using for your
codebase. This should allow Mypy to do a better job of detecting what version
of Python to use. In addition, in the future, checks will be able to grab
the passed in version and give different error messages for the version to
be used. For example, using type unions in certain areas (such as the second
arg to `isinstance()`).