Fix `--load` being able to load a single check multiple times:
When loading the checks, we did not check that to see if that specific module
had been loaded already, meaning you could the same check twice on accident.
This also applies to nested modules. For example, loading the module "a.b.c"
does nothing if "a" or "a.b" has already been loaded.