Allow for dependency injecting settings into checks:
Checks can now request access to the settings environment used for running
Refurb. The most useful purpose for this will checking the version of Python
as supplied by the user. For example, FURB121 now reccomends using a type union
instead of a tuple for calls to `isinstance()` and `issubclass()`, but only
in Python 3.10 and up.

As more dependency injectable services are added, we will need to change how
these dependencies are loaded. Currently the `__annotations__` attribute is
checked to see how many parameters there are, but this won't work once more
dependencies are made available. Also, this is not very type safe, and prone
to error. A possible solution would be using decorators, though doing this in
a type safe way will be interesting: We will either have to reinspect the
check function before we run it, or store the information in a `CheckWrapper`
type of sorts.