Improve and disable the `expandtabs()` check:
I initially wanted to add support for parsing `n * " "` expressions as the
second argument to `.replace()`, but found out that the `expandtabs()` function
doesn't do what I thought it did. Because of this, I am disabling it, since it
is encouraging changes which (could) alter the behaviour of the program.