Add better node equivalence checks (#98):
Previously, nodes where deemed to be similar if their string representations
where the same. This caused issues, since the string representations contain
line number (sometimes), which would cause an error to sometimes not be emitted
if a similar node was on a different line.

Now the more common nodes (NameExpr and MemberExpr) are properly checked,
and the string comparison is used as a fallback. This is by no means an
exhaustive fix, but at least fixes the most common use-cases (that I can think
of).

Closes #97.