Fix mypy issue:
This is the second time recently that mypy has passed on my machine, but failed
on the CI server. When I run `mypy test/test_category_naming.py` directly, it
fails, but when I run `mypy -p test` it does not. This is probably some sort of
caching issue, but I will have to dig into this at a later time.