Add "use comprehensions" check:
This check will find use cases where you create an empty list, append every
element of some iterable in a for loop to this new list, and nothing else.
There are a lot of edge cases that we have to exclude, which is better then
the alternative (false positives which result in just outright disabling this
check).

I have spent enough time on this already, but I would like to add set and
dict comprehensions as well since the semantics are pretty much the same.