Fix set comprehensions being suggested to be removed for list/tuple constructions (#131):
Because sets cannot contain unique elements, you can use a set comprehension to
eliminate any duplicates, and then convert the result to a list or tuple. This
would change though if you dropped the set comprehension, because the length of
the resulting list would contain more elements, and would not be unique.