Fix mypy errors:
This is the 5th time this has happened, so I decided to take the time to look
into this. I do not know why, but removing the "-p" flag seems to cause the
errors to pop up again. What is curious is that there is no `common.meta.json`
or `common.data.json` file in the `.mypy_cache` folder, perhaps that is why the
error was not appearing when the `-p` flag was passed? If that is the case,
why wouldn't mypy recreate it?

Regardless, this (hopefully) fixes my issue for now. I will open a PR for this
once I am able to widdle it down to a more exact cause.