Add "no set in for loop" check (#140):
This check only checks for "add" and "discard" being called on a set in a for
loop, since those are the only functions which have an in-place modifying
equivalent. The "add" function in a for loop will interfere with the eventual
"use set comprehension" check, assuming that you are assigning to a newly
created set. I don't know what the best way to handle this is. The same probably
applies for calling "append" in a for loop. The best solution in this case is
to just let 2 errors get emitted, and let the user use their discretion to
choose what suits their needs best.