Fix `.pyi` files taking precedence over `.py` files:
From the mypy docs (https://mypy.readthedocs.io/en/stable/stubs.html):

> If a directory contains both a `.py` and a `.pyi` file for the same module,
> the `.pyi` file takes precedence.

This is an issue because the actual source code we care about is not in the
`.pyi` files, it is in the `.py` files. I would assume the amount of type
information is less in these files, and so certain errors might not work
properly, though this is much better then ignoreing the file entirely.

If this becomes an issue where people need the `.pyi` files, we will have to
add a flag to disable this behaviour. The only reason I can see where this will
cause an issue is if you run Refurb on a file with only `.pyi` files, and
since they are all ignored, you will get an error.