Add `noqa` comments to test files (#148):
To make it clearer as to what is actually being tested, I updated the test
files to only error on the error codes for that particular file.

In the future, I plan on leveraging pytests custom test cases, which basically
allows you to build custom test cases at runtime. Currently, all the test data
files are compiled into one big list, and if one item in the list is misplaced,
the test fails (and it is very hard to see what changed). After updating to the
test case method, there would be a separate test for each error code.