Fix FURB135 false positives (#153):
Previously this check only checked if the unpacked tuple variables where used
in the body of the for loop. This is an issue because the variable can still
be accessed after the for loop is done executing, and currently there is no
way to check if this is the case.

On the flip side, this check now checks for unused variables in comprehension
expressions, since the unpacked values in those expressions are ran in their
own context, meaning you can check whether they are used or not much easier.