Add "no is bool compare" check:
This checks for comparing a boolean variable to `True` or `False` using an `is`
or `is not` operator. This is just a more verbose way of doing a truthy/falsey
check, and can be simplified because the type is a simple boolean.

Currently this only checks for boolean variables, not member expressions such
as `x.y`, function calls, etc. This will be expanded on in the future.