Add `name` field to checks:
This is merely a placeholder which future PR's will build off of. Once this
feature is fully complete, you will be able to use the check "name" in `noqa`
comments, in the CLI, and inconfig files. Finally, the name of the check (along
with the categories) will be used for displaying when using the `--explain`
flag.