Add ability to ignore errors on a per file/folder basis:
Closes #29.

Now you can ignore errors by id and category for a specific file or folder.
This is only allowed in the config file, and uses TOML tables/arrays to specify
which ignores should apply to which paths. All paths are relative to the config
file.

Read the "Ignore Checks Per File/Folder" section of the README for more info
on how to use this new feature.