Remove part of FURB131 (#239):
This PR removes a portion of FURB131 which suggested replacing most instances
of `del` with `.pop()`. The reasoning behid it was not well founded, and in
most cases, it is better to just leave the `del` statememt as it is. In one
instance though, which is the `del x[:]` usage, it is more preferable to use
the `.clear()` method, as it is faster and more readable.

Closes #234.