Add `rstrip` and `strip` support to FURB139 (#240):
This PR detects multi-line strings used with `rstrip()` and `strip()`. The
error message reflects what changes need to be made based off of whether there
are leading/trailing newlines in the string. In addition, a bug was fixed where
if a newline literal was passed to the `strip()` function, it would not show
the argument in the error message.