Improve coverage:
Closes #261.

I ended up not fully enabling branch coverage because most of the missing
branch coverage lines where false positives. For example, Refurb uses a custom
type-registration system where check functions declare the types that they
want to respond to, and Refurb will call the check when that node type is
present. This means it is impossible for other types to be passed at run time.
The coverage system did not like this though, and expected me to explicitly
add exhaustive clauses to match statements to verify that it shouldn't be
allowed to get here. If there was a way do disable certain features of the
branch coverage system I would be more inclined to enable it, but right now it
is too strict.