Add `--timing-stats` flag:
This will hopefully help me address some of the slowness in Refurb. There is a
lot of area for improvement, but right now it is hard/impossible to pinpoint
where the slowness is. I will probably add more datapoints in the future, but
for now this should be sufficient to get some preliminary timing stats.

Also bump package version.