Fix tests failing in Python 3.12:
For whatever reason I cannot access `__dict__` on my `RefurbVisitor` class
anymore in Python 3.12. I should look into why at some point (for curiosity),
but for this will work. There is probably a prettier way of doing this, but
because it is a test file that doesn't get changed/viewed often I'm fine with
it the way it is for now.