Skip files that raise `RecursionError`:
Certain deeply-nested expressions cause the Refurb visitor to blow the stack.
This doesn't seem to be an issue when running with Mypy directly, though this
is probably because we are traversing the entire AST. As a workaround we will
silently ignore files that cause this issue. For debugging purposes there
should probably be a flag to show files that have been ignored (perhaps just
stuff it under the `-v` flag for now).

Also bump version for new release.

Closes #302.