Pin Mypy to <1.7.0 (#306):
Temporarily fixes #305.

Specifically, this error is being caused by this line:

https://github.com/python/mypy/blob/e4355948d797600c7b76da0a916fc5f29d10448e/mypy/traverser.py#L97

Because `TraverserVisitor` now is a `trait` you cannot extend from it. There
doesn't seem to be a way to get around this on my end, so I'm pinning Mypy as a
workaround until I find a solution or something changes upstream with Mypy.