Fix syntax errors in docstring code snippets:
Some of the code examples in the docstrings where not valid Python code, so I
wrote a small test to ensure all the checks have valid Python code. This
doesn't do anything fancy like checking for imports (though maybe it should),
it merely checks that the AST tree can be built.