Add more cases to FURB111:
FURB111 now detects instances of `lambda: x`, where `x` is the default literal
value for a builtin type. For example, the default value of `bool()` is
`False`, `int()` is `0`, and so on. Similar to how `lambda: []` should be
written as `list`, instances of `lambda: 0` (and other literal values) should
be written as `int`.

Pyupgrade can already detect these new cases, but for whatever reason, it only
detects them in `defaultdict()` calls and nowhere else. With that in mind I
decided to implement this anyways, but without the `defaultdict()` restriction.
Pyupgrade also doesn't check `lambda: b""`, though this is probably not very
common.