Reimplement FURB131:
FURB131 (previously `no-del`) has been broken for a while, basically since it's
initial creation. #239 fixed some of this, but introduced improper semantics by
implying you could use `del x[:]` when `x` is a `dict`, which is not true. This
means that FURB131 only really applies to `list` types now (or any sequence
that implements `.clear()`, though this isn't supported yet). I recently came
across a similar idiom for clearing an array, `x[:] = []`, which is about twice
as slow as `x.clear()`, and similar in nature to `del x[:]`. This new update
was what prompted me to fix all the broken logic.

With that in mind, this check has been renamed to `use-clear` since it no
longer is about the `del` statement, but about using slices to clear a list.
This shouldn't cause any issues since the name of checks are only used for
display purposes. This will break any hyperlinks to FURB131 in the docs, though
you'll still end up on the same page so I'm not too concerned about that.