Add better type deduction:
This part one of many to add better type deduction to Refurb. Previously I only
pulled type out of `NameExpr` nodes because it was convienent, but this left
out a lot of node types, meaning Refurb could not detect them. This new change
allows for extracting the type of an AST node, meaning Refurb will (in the
future) be much better at detecting certain checks.

This is still in the works and has only been applied to one check, but as time
goes on, this will start to be implemented for more checks.