Improve type deduction of string literals:
This commit adds support for detecting string literal expressions in the new
`get_mypy_type()` function. To get this to work properly we need to store a
copy of the `builtins` module as a global variable so we can extract complete
type information for types like `str`. There might be a better way to do this,
but this works for the time being.

Also improve error messages for FURB159 since I was using this check to test
these new changes.