Bump dev dependencies:
The code fixes where from enabling preview mode in Black. I merged a PR that
removed redundant parens in match guards, but this seems to be acting weird
under certain circumstances and actually adding parens where it shouldn't. I
will have to investigate why this is happening, but for now I only applied the
changes that it was supposed to catch.