Type deduce object subclasses:
Closes #329.

This commit adds the ability to detect the subclass of a given class, allowing
Refurb to safely detect types derived from ABC classes like `Sequence`,
`Mapping`, and so on.

This commit will give more warnings in user code, and even gave a warning for
a line in Refurb.

There are still many improvements to the type system left to implement, but
this was the big one on my list.