Disallow Mypy version 1.11.0:
See #347

Mypy 1.11.0 was causing crashes when used with Refurb, but was resolved with
the release of Mypy 1.11.1. This commit prevents Mypy version 1.11.0 from
being installed alongside Refurb. While this issue is old and likely not an
issue anymore, it is still good to indicate that this version is incompatible
with Refurb.