Disallow nested function calls in FURB111:
Closes #343

While code like `lambda: f()` can be re-written as just `f`, code like
`lambda: f().g()` cannot be re-written as `f().g`, as `f()` would be
eagerly evaluated, but in the lambda version it is lazily evaluated.

There are probably other edge cases I am missing, but this fixes the immediate
issue. In theory, lots of cases might be "unsafe" here, but this is a very
common, easily detectable scenario.