Fix failing tests, require Mypy 1.10 or higher:
Since type alias support was added to Mypy 1.10, we have to require that
version in refurb in order for things to work properly. There *might* be a
way to gracefully fallback on other versions, but since the 1.10 release is
over a year old, I don't suspect this will be a big issue.